+++ src/app/signup/_component/funnel/SelectCategory.tsx
@@ -3,7 +3,7 @@
 import { Chips } from "@/app/_component/common/Chips";
 import { Chip } from "@/app/_component/common/Chips";
 
-import useOnboardingStore from "../store/store";
+import useOnboardingStore from "../../store/store";
 
 const SelectCategory = () => {
   const category = useOnboardingStore((state) => state.category);
