+++ src/app/signup/_component/funnel/SelectResidence.tsx
